added privacy score checks and some tests, fixed it in app service

+++ api/models.go
@@ -394,13 +394,13 @@ type ToolInput struct {
 	Data struct {
 		Type       string `json:"type"`
 		Attributes struct {
-			Name           string  `json:"name"`
-			Description    string  `json:"description"`
-			ToolType       string  `json:"tool_type"`
-			OASSpec        []byte  `json:"oas_spec"`
-			PrivacyScore   float64 `json:"privacy_score"`
-			AuthKey        string  `json:"auth_key"`
-			AuthSchemaName string  `json:"auth_schema_name"`
+			Name           string `json:"name"`
+			Description    string `json:"description"`
+			ToolType       string `json:"tool_type"`
+			OASSpec        []byte `json:"oas_spec"`
+			PrivacyScore   int    `json:"privacy_score"`
+			AuthKey        string `json:"auth_key"`
+			AuthSchemaName string `json:"auth_schema_name"`
 		} `json:"attributes"`
 	} `json:"data"`
 }
@@ -415,7 +415,7 @@ type ToolResponse struct {
 		Description    string   `json:"description"`
 		ToolType       string   `json:"tool_type"`
 		OASSpec        []byte   `json:"oas_spec"`
-		PrivacyScore   float64  `json:"privacy_score"`
+		PrivacyScore   int      `json:"privacy_score"`
 		Operations     []string `json:"operations"`
 		AuthKey        string   `json:"auth_key"`
 		AuthSchemaName string   `json:"auth_schema_name"`
