+++ api/tool_handlers.go
@@ -320,7 +320,7 @@ func serializeTool(tool *models.Tool) ToolResponse {
 			Description    string   `json:"description"`
 			ToolType       string   `json:"tool_type"`
 			OASSpec        []byte   `json:"oas_spec"`
-			PrivacyScore   float64  `json:"privacy_score"`
+			PrivacyScore   int      `json:"privacy_score"`
 			Operations     []string `json:"operations"`
 			AuthKey        string   `json:"auth_key"`
 			AuthSchemaName string   `json:"auth_schema_name"`
