+++ services/tool_service.go
@@ -5,7 +5,7 @@ import (
 )
 
 // CreateTool creates a new tool with validity checks
-func (s *Service) CreateTool(name, description, toolType string, oasSpec []byte, privacyScore float64) (*models.Tool, error) {
+func (s *Service) CreateTool(name, description, toolType string, oasSpec []byte, privacyScore int) (*models.Tool, error) {
 	tool := &models.Tool{
 		Name:         name,
 		Description:  description,
@@ -22,7 +22,7 @@ func (s *Service) CreateTool(name, description, toolType string, oasSpec []byte,
 }
 
 // UpdateTool updates an existing tool with validity checks
-func (s *Service) UpdateTool(id uint, name, description, toolType string, oasSpec []byte, privacyScore float64) (*models.Tool, error) {
+func (s *Service) UpdateTool(id uint, name, description, toolType string, oasSpec []byte, privacyScore int) (*models.Tool, error) {
 	tool, err := s.GetToolByID(id)
 	if err != nil {
 		return nil, err
